// ballroom-app/src/app/ballrooms/models/contact.ts
export class Contact {
    id: number;
    telephone: string;
    address: string;
    email: string;
    site: string;
    ballroomId: string[];
    constructor(
        id: string,
        telephone: string,
        address: string,
        email: string,
        site: string,
        ballroomId: string[] ) { }

}